Apply U-Turn to peds as well
