+++ ChaosMod/Effects/EffectsInfo.h
@@ -294,7 +294,7 @@ enum EEffectType : int
 	EFFECT_MISC_STUFFGUNS,
 	EFFECT_PEDS_ETERNAL_SCREAMS,
 	EFFECT_PLAYER_BLADE_HUNGER,
-	EFFECT_VEHS_UTURN,
+	EFFECT_MISC_UTURN,
 	_EFFECT_ENUM_MAX
 };
